+++ src/sql/__init__.py
@@ -1,5 +1,4 @@
 from .magic import RenderMagic, SqlMagic, load_ipython_extension
-from .error_message import SYNTAX_ERROR
 from .connection import PLOOMBER_DOCS_LINK_STR
 
 __version__ = "0.7.10dev"
@@ -9,6 +8,5 @@ __all__ = [
     "RenderMagic",
     "SqlMagic",
     "load_ipython_extension",
-    "SYNTAX_ERROR",
     "PLOOMBER_DOCS_LINK_STR",
 ]