+++ src/sql/command.py
@@ -71,6 +71,10 @@ class SQLCommand:
         if add_alias:
             self.parsed["connection"] = self.args.line[0]
 
+        if self.args.with_:
+            final = store.render(self.parsed["sql"], with_=self.args.with_)
+            self.parsed["sql"] = str(final)
+
     @property
     def sql(self):
         """