+++ src/tests/test_magic.py
@@ -1235,16 +1235,10 @@ def test_save_with_number_table(
 
 
 def test_save_with_non_existing_with(ip):
-    with pytest.warns(FutureWarning) as record:
-        ip.run_cell(
-            "%sql --with non_existing_sub_query " "SELECT * FROM non_existing_sub_query"
-        )
-    assert len(record) == 1
-    assert (
-        "CTE dependencies are now automatically inferred, you can omit the "
-        "--with arguments. Using --with will raise an exception in the next "
-        "major release so please remove it." in record[0].message.args[0]
+    out = ip.run_cell(
+        "%sql --with non_existing_sub_query " "SELECT * FROM non_existing_sub_query"
     )
+    assert isinstance(out.error_in_exec, UsageError)
 
 
 def test_save_with_non_existing_table(ip, capsys):