+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [Fix] Refactored `ResultSet` to lazy loading (#470)
 * [Fix] Removed `WITH` when a snippet does not have a dependency (#657)
 * [Fix] Used display module when generating CTE (#649)
+* [Fix] Adding `--with` back because of issues with sqlglot query parser (#684)
 
 ## 0.7.9 (2023-06-19)